requirements.txt
ttkbootstrap
opencv-contrib-python
numpy
scikit-image
scipy
ultralytics
Pillow==9.5.0